+++ .gitignore
@@ -22,4 +22,4 @@ npm-debug.log*
 yarn-debug.log*
 yarn-error.log*
 
-mapbox-accessToken.js
+./src/mapbox-accessToken.js
