+++ src/mapbox-accessToken.js
@@ -1,0 +1,3 @@
+const accessToken = 'pk.eyJ1IjoiYWl0aG9uMjUwMSIsImEiOiJjanMxODMxcnkxcmphNDRvMzM2MWt2cTd3In0.5yin38YovLGgVUHnRo9FjA';
+
+export default accessToken;
